+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - [#73](https://github.com/Kashoo/synctos/issues/73): Include an implicit type property when a simple type filter is used
+- [#78](https://github.com/Kashoo/synctos/issues/78): Enum property validation type
 
 ## [1.6.0] - 2017-01-18
 ### Added